Refactored and re organised fixtures and completed the phone auth
- linking phone numnber to current account
- datasets introduced
