--- src/components/PhoneAuth/index.tsx
@@ -1,5 +1,0 @@
-import React from "react";
-
-export const PhoneAuth: React.FC = () => {
-  return null;
-};
