+++ src/constant/datasets/defaults/authentication.ts
@@ -1,0 +1,28 @@
+const data = {
+  "login.subtext": "Please login to continue with your account",
+  "login.form.title": "Sign in",
+  "login.form.email-input.placeholder": "Email address",
+  "login.form.password-input.placeholder": "Password",
+  "login.form.submit.buttontext": "Login",
+  "login.form.register.buttontext": "Create new account",
+  "register.subtext": "Register with your new email adress",
+  "register.form.title": "Create new account",
+  "register.form.email-input.placeholder": "Email address",
+  "register.form.password-input.placholder": "Password",
+  "register.form.name-input.placeholder": "Name",
+  "register.form.submit.buttontext": "Create account",
+  "register.form.login.buttontext": "Or login here",
+  "email-confirmation.pre.title": "Confirm your email address",
+  "email-confirmation.post.title": "Email Sent",
+  "email-confirmation.pre.description":
+    "You have to verify and confirm your Email address before login into Napigo",
+  "email-confirmation.post.description":
+    "We've sent a link to your email address for confirmation. Please check your email and click on the link provided.",
+  "email-confirmation.idle.buttontext": "Send confirmation link",
+  "email-confirmation.logoutnow.buttontext": "Logout now",
+  "email-confirmation.onsuccess.buttontext": "Resend Link",
+  "email-confirmation.loading.buttontext": "Sending ...",
+  "email-confirmation.onerror.buttontext": "Sending failed",
+};
+
+export default data;
