+++ src/constant/datasets/defaults/commons.ts
@@ -1,0 +1,5 @@
+const data = {
+  "alert.error.title": "Error!",
+};
+
+export default data;
