+++ src/constant/datasets/defaults/loadings.ts
@@ -1,0 +1,5 @@
+const data = {
+  "apploader.text": "",
+};
+
+export default data;
