+++ src/constant/datasets/defaults/nav.ts
@@ -1,0 +1,11 @@
+const data = {
+  "service.buttontext": "Services",
+  "search.placeholder": "Search anything",
+  "usermenu.dropdown.title": "User Menu",
+  "usermenu.dropdown.item.setting.linktext": "Settings",
+  "usermenu.dropdown.item.profile.linktext": "Profile",
+  "usermenu.dropdown.item.members.linktext": "Members",
+  "usermenu.dropdown.item.logout.linktext": "Logout",
+};
+
+export default data;
