+++ src/constant/datasets/defaults/settings.ts
@@ -1,0 +1,56 @@
+const data = {
+  "settings-menu.general.buttontext": "General",
+  "settings-menu.authentication.buttontext": "Authentication",
+  "authentication.title": "Authentication",
+  "authentication.back-console.buttontext": "Back to console",
+  "authentication.emailloginform.heading": "Login Email Address",
+  "authentication.emailloginform.primary.input.label": "Primary",
+  "authentication.emailloginform.primary.input.helpertext":
+    "This is your default email which is used to login to your account",
+  "authentication.emailloginform.alternative.input.label": "Alternative Email",
+  "authentication.emailloginform.alternative.input.placeholder":
+    "Enter Alternative Email Address",
+  "authentication.emailloginform.submit.buttontext": "Confirm",
+  "authentication.emailloginform.cancel.buttontext": "Cancel",
+  "authentication.emailloginform.add.buttontext": "Cancel",
+  "authentication.passwordform.heading": "Password -  Credential",
+  "authentication.passwordform.readonly-input": "************",
+  "authentication.passwordform.changepassword.buttontext": "Change password",
+  "authentication.passwordform.forgotpassword.buttontext": "Forgot password ?",
+  "authentication.passwordform.input.currentpassword.label": "Current password",
+  "authentication.passwordform.input.newpassword.label": "New password",
+  "authentication.passwordform.input.newpassword.helpertext":
+    "Please make sure that your password is at least 8 charaters with alphanumeric and digits",
+  "authentication.passwordform.submit.buttontext": "Confirm",
+  "authentication.passwordform.cancel.buttontext": "Cancel",
+  "authentication.mobileform.heading": "Mobile Number",
+  "authentication.mobileform.verified.description": `Your phone number can be used as backup when losing your
+  password or forgotten your login email address. It can be used
+  for MFA, please refer to the MFA section on setup.`,
+  "authentication.mobileform.notverified.description": `Provide us with your mobile number to allow mobile
+  authentication for your login method or you can use for backup
+  in case you couldnt logon to your email address`,
+  "authentication.mobileform.verified-alert.title": "Verified!",
+  "authentication.mobileform.verified-alert.description": `Your Phone number have been added, you can now setup MFA for
+  better security login`,
+  "authentication.mobileform.readonly.input.placeholder": "Mobile number",
+  "authentication.mobileform.update-number.buttontext": "Update phone number",
+  "authentication.mobileform.notverified.select-country.input.placeholder":
+    "Select country",
+  "authentication.mobileform.notverified.mobile-no.input.placeholder":
+    "Mobile number",
+  "authentication.mobileform.notverified.send-otp.buttontext": "Send OTP",
+  "authentication.mobileform.update-no.cancel.buttontext": "Cancel",
+  "authentication.mobileform.otp-sent.description":
+    "Enter the code we`ve just sent on your mobile phone \n",
+  "authentication.mobileform.otp-success.title": "Verified!",
+  "authentication.mobileform.otp-success.description": `Your Phone number have been added, you can now setup MFA for
+  better security login`,
+  "authentication.mobileform.otp-resent.description":
+    "Re sent OTP to your mobile!",
+  "authentication.mobileform.otp-resend.buttontext": "Resend OTP",
+  "authentication.mobileform.otp.use-another-number.buttontext":
+    "Use another number",
+};
+
+export default data;
