+++ src/containers/Settings/index.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { VerticalMenuLayout } from "@/layouts/VerticalMenuLayout";
 import { Navigate, Route, Routes } from "react-router-dom";
-import { SettingsMenu } from "@/components/SettingsMenu";
+import { SettingsMenu } from "@/components/Settings/SettingsMenu";
 import { SettingGeneralPage } from "@/pages/Settings/General";
 import { SettingAuthenticationPage } from "@/pages/Settings/Authentication";
