+++ src/layouts/UserLayout.tsx
@@ -3,8 +3,6 @@ import { Outlet } from "react-router-dom";
 import { Nav } from "@/components/Nav";
 
 export const UserLayout: React.FC = () => {
-  // const [drawerCollapse, setDrawerCollapse] = useState<boolean>(false);
-
   return (
     <Fragment>
       <Nav />
